requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aioice==0.9.0
aiortc==1.11.0
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
anyio==3.7.1
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
attrs==25.3.0
av==14.2.0
beautifulsoup4==4.12.3
billiard==4.2.1
bitarray==3.1.1
black==24.10.0
Brotli==1.1.0
cachetools==5.5.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
channels==4.1.0
channels_redis==4.2.1
chardet==3.0.4
charset-normalizer==3.4.0
ckzg==2.1.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
crispy-bootstrap4==2024.10
crispy-bootstrap5==2024.10
cryptography==44.0.2
cssselect2==0.7.0
cytoolz==1.0.1
deep-translator==1.11.4
defusedxml==0.7.1
distro==1.9.0
Django==5.1.2
django-allauth==65.1.0
django-bootstrap4==24.4
django-channels==0.7.0
django-cors-headers==4.7.0
django-countries==7.6.1
django-crispy-forms==2.3
django-extensions==3.2.3
django-filter==24.3
django-guardian==2.4.0
django-model-utils==5.0.0
django-modeltranslation==0.19.10
django-redis==5.4.0
django-sslserver==0.22
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
dnspython==2.7.0
eth-account==0.13.5
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.6.1
eth-rlp==2.2.0
eth-typing==5.2.0
eth-utils==5.2.0
eth_abi==5.2.0
factory_boy==3.3.1
Faker==30.8.0
fonttools==4.55.1
frozenlist==1.5.0
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-core==2.4.1
google-cloud-translate==3.19.0
google-crc32c==1.7.1
googleapis-common-protos==1.66.0
googletrans==4.0.0rc1
grpc-google-iam-v1==0.13.1
grpcio==1.68.1
grpcio-status==1.68.1
gunicorn==23.0.0
h11==0.9.0
h2==3.2.0
hexbytes==1.3.0
hpack==3.0.0
hstspreload==2024.12.1
html5lib==1.1
httpcore==0.9.1
httplib2==0.22.0
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
ifaddr==0.2.0
Jinja2==3.1.4
jiter==0.8.2
jwt==1.3.1
kombu==5.4.2
lxml==5.3.0
MarkupSafe==3.0.2
msgpack==1.1.0
multidict==6.2.0
mypy-extensions==1.0.0
oauthlib==3.2.2
oscrypto==1.3.0
packaging==24.1
parsimonious==0.10.0
pathspec==0.12.1
paystackapi==2.1.3
phonenumbers==8.13.52
pillow==10.4.0
platformdirs==4.3.6
polib==1.2.0
prompt_toolkit==3.0.48
propcache==0.3.0
proto-plus==1.25.0
protobuf==5.29.1
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.10.3
pydantic_core==2.27.1
pydyf==0.11.0
pyee==13.0.0
pyHanko==0.25.1
pyhanko-certvalidator==0.26.3
PyJWT==2.10.1
pylibsrtp==0.11.0
pyOpenSSL==25.0.0
pyparsing==3.2.0
pypdf==5.0.1
pyphen==0.15.0
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-decouple==3.8
python3-openid==3.2.0
pytz==2024.2
pyunormalize==16.0.0
PyYAML==6.0.2
qrcode==8.0
redis==5.2.0
regex==2024.11.6
reportlab==4.2.5
requests==2.32.3
requests-oauthlib==2.0.0
rfc3986==1.5.0
rlp==4.1.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.4.3
social-auth-core==4.5.6
soupsieve==2.6
sqlparse==0.5.1
stripe==11.5.0
svglib==1.5.1
tinycss2==1.4.0
tinyhtml5==2.0.0
toolz==1.0.0
tqdm==4.67.1
types-requests==2.32.0.20250306
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
uritemplate==4.1.1
uritools==4.0.3
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
web3==7.9.0
webencodings==0.5.1
websockets==13.1
whitenoise==6.9.0
xhtml2pdf==0.2.16
yarl==1.18.3
zopfli==0.2.3.post1
